Add CraftMultiInventory
CraftMultiInventory is meant to be a base class for inventories such as
CraftInventoryAnvil and CraftInventoryCrafting, which are composed of
multiple NMS inventories.

I have not made the changes to CraftInventoryCrafting and
CraftInventoryAnvil as this is meant to simply demonstrate the concept
before actually making the changes.

A potentially better solution would be to simply make these changes to
CraftInventory.  That would make CraftInventory a more suitable base class
for any implementation of Inventory.